+++ Estimation/CellsDataContainer.h
@@ -3,6 +3,7 @@
 #include "Cell.h"
 #include "Stats.h"
 #include "Tools/IndexedValue.h"
+#include <Tools/GtfRecord.h>
 #include "UMI.h"
 
 #include <string>
@@ -10,7 +11,7 @@
 #include <map>
 #include <vector>
 #include <unordered_map>
-#include <Tools/GtfRecord.h>
+#include <fstream>
 
 namespace TestEstimator
 {
@@ -37,10 +38,11 @@ namespace Estimation
 	public:
 		typedef std::unordered_map<std::string, std::string> s_s_hash_t;
 		typedef std::unordered_map<std::string, size_t> s_ul_hash_t;
+		typedef std::unordered_map<std::string, int> s_i_hash_t; // not long because of RCpp
 
 		typedef std::vector<Tools::IndexedValue> i_counter_t;
 		typedef std::vector<size_t> ids_t;
-		typedef std::vector<size_t> counts_t;
+		typedef std::vector<int> counts_t;
 		typedef std::vector<std::string> names_t;
 
 	private:
@@ -50,12 +52,12 @@ namespace Estimation
 		static const size_t TOP_PRINT_SIZE;
 		const int _max_cells_num;
 
-		std::vector<Cell> _cells; //cell_id -> gen_name -> umi -> count
+		std::vector<Cell> _cells; //cell_id -> gen_name -> umi -> #reads
 		s_ul_hash_t _cell_ids_by_cb;
-		ids_t _filtered_cells;
+		ids_t _filtered_cells; // Sorted ascending by the number of genes
 		ids_t _merge_targets;
 
-		i_counter_t _filtered_gene_counts_sorted;
+//		i_counter_t _filtered_gene_counts_sorted;
 		bool _is_initialized;
 		const std::vector<UMI::Mark> _query_marks;
 
@@ -64,19 +66,18 @@ namespace Estimation
 		size_t _has_not_annotated_reads;
 		size_t _number_of_real_cells;
 
-		mutable Stats _stats;
-
 	private:
-		std::string get_cb_count_top_verbose(const i_counter_t &cells_genes_counts) const;
+		std::string get_cb_count_top_verbose() const;
 		size_t update_cell_sizes(size_t requested_genes_threshold, int cell_threshold);
+		void update_cell_stats(size_t cell_id, const UMI::Mark &mark, const std::string &chromosome_name);
 
 	public:
 		CellsDataContainer(std::shared_ptr<Merge::MergeStrategyAbstract> merge_strategy,
 		                   std::shared_ptr<Merge::UMIs::MergeUMIsStrategySimple> umi_merge_strategy,
 		                   const std::vector<UMI::Mark> &gene_match_levels, int max_cells_num = -1);
 
-		size_t add_record(const std::string &cell_barcode, const std::string &umi, const std::string &gene,
-		                  const UMI::Mark &umi_mark = UMI::Mark::HAS_EXONS);
+		void add_record(const std::string &cell_barcode, const std::string &umi, const std::string &gene,
+		                const std::string &chr_name = "", const UMI::Mark &umi_mark = UMI::Mark::HAS_EXONS);
 		void exclude_cell(size_t index);
 
 		void merge_and_filter();
@@ -87,23 +88,25 @@ namespace Estimation
 		void set_initialized();
 
 		size_t total_cells_number() const;
-		const i_counter_t& cells_gene_counts_sorted() const;
 		const s_ul_hash_t& cell_ids_by_cb() const;
-		names_t excluded_cells() const;
 		const ids_t& filtered_cells() const;
 		const ids_t& merge_targets() const;
 		const std::vector<UMI::Mark>& gene_match_level() const;
-		size_t get_filtered_gene_counts(size_t requested_genes_threshold, int cell_threshold,
-		                                i_counter_t& filtered_gene_counts_sorted) const;
+		size_t update_filtered_gene_counts(size_t requested_genes_threshold, int cell_threshold);
+		s_i_hash_t get_stat_by_real_cells(Stats::CellStatType type) const;
+		void get_stat_by_real_cells(Stats::CellChrStatType stat, names_t &cell_barcodes, names_t &chromosome_names,
+		                            counts_t &counts) const;
+
+		bool compare_cells(size_t cell1_id, size_t cell2_id) const;
 
 		const Cell &cell(size_t index) const;
+		Cell &cell(size_t index);
 
 		size_t has_exon_reads_num() const;
 		size_t has_intron_reads_num() const;
 		size_t has_not_annotated_reads_num() const;
 		size_t real_cells_number() const;
 
-		Stats &stats() const;
 		std::string merge_type() const;
 
 		s_ul_hash_t umis_distribution() const;