+++ Estimation/Gene.h
@@ -30,7 +30,7 @@ namespace Estimation
 		s_ul_hash_t requested_reads_per_umi(const UMI::Mark::query_t &query) const;
 
 		bool add_umi(const std::string &umi, const UMI::Mark &umi_mark);
-		size_t merge(const Gene& source);
+		void merge(const Gene& source);
 		void merge(const std::string& source_umi, const std::string& target_umi);
 	};
 }