+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Upcoming]
+### Changed
+* Optimized memory usage for dropEst
+* Using of stable sort for cells selection (by number of genes)
+
 ## [0.6.1] - 2017-09-05
 ### Added
 * Support for pseudoaligners .bam format (usage of chromosome name as a source of gene name)